refactor: move range methods on Cfg to graph.rs
