--- riscv_analysis/src/lints/checks.rs
@@ -1,98 +1,0 @@
-use crate::analysis::HasGenKillInfo;
-use crate::cfg::Cfg;
-use crate::cfg::CfgNode;
-use crate::parser::InstructionProperties;
-use crate::parser::Register;
-use crate::parser::RegisterToken;
-use std::collections::{HashSet, VecDeque};
-use std::rc::Rc;
-
-// If we need to add an error to a register at its first use/store, we need to
-// know their ranges. This function will take a register and return the ranges
-// that need to be annotated. If it cannot find any, then it will return the original
-// node's range.
-impl Cfg {
-    /// Perform a backwards search to find the first node that stores to the given register.
-    ///
-    /// From a given end point, like a return value, find the first node that stores to the given register.
-    /// This function works by traversing the previous nodes until it finds a node that stores to the given register.
-    /// This is used to correctly mark up the first store to a register that might
-    /// have been incorrect.
-    pub fn error_ranges_for_first_store(node: &Rc<CfgNode>, item: Register) -> Vec<RegisterToken> {
-        let mut queue = VecDeque::new();
-        let mut ranges = Vec::new();
-        // push the previous nodes onto the queue
-        queue.extend(node.prevs().clone());
-
-        // keep track of visited nodes
-        #[allow(clippy::mutable_key_type)]
-        let mut visited = HashSet::new();
-        visited.insert(Rc::clone(node));
-
-        // visit each node in the queue
-        // if the error is found, add error
-        // if not, add the previous nodes to the queue
-        while let Some(prev) = queue.pop_front() {
-            if visited.contains(&prev) {
-                continue;
-            }
-            visited.insert(Rc::clone(&prev));
-            if let Some(reg) = prev.writes_to() {
-                if *reg.get() == item {
-                    ranges.push(reg);
-                    continue;
-                }
-            }
-            queue.extend(prev.prevs().clone().into_iter());
-        }
-        ranges
-    }
-
-    // TODO move to a more appropriate place
-    // TODO make better, what even is this?
-    pub fn error_ranges_for_first_usage(node: &Rc<CfgNode>, item: Register) -> Vec<RegisterToken> {
-        let mut queue = VecDeque::new();
-        let mut ranges = Vec::new();
-        // push the next nodes onto the queue
-
-        queue.extend(node.nexts().clone());
-
-        // keep track of visited nodes
-        #[allow(clippy::mutable_key_type)]
-        let mut visited = HashSet::new();
-        visited.insert(Rc::clone(node));
-
-        // visit each node in the queue
-        // if the error is found, add error
-        // if not, add the next nodes to the queue
-        while let Some(next) = queue.pop_front() {
-            if visited.contains(&next) {
-                continue;
-            }
-            visited.insert(Rc::clone(&next));
-            if next.gen_reg().contains(&item) {
-                // find the use
-                let regs = next.reads_from();
-                let mut it = None;
-                for reg in regs {
-                    if reg == item {
-                        it = Some(reg);
-                        break;
-                    }
-                }
-                if let Some(reg) = it {
-                    ranges.push(reg);
-                    break;
-                }
-                break;
-            }
-
-            queue.extend(next.nexts().clone().into_iter());
-        }
-        ranges
-    }
-}
-
-// Checks are passes that occur after the CFG is built. As much data as possible is collected
-// during the CFG build. Then, the data is applied via a check.
-// TODO check if the stack is ever stored at 0 or what not
