+++ riscv_analysis/src/lints/mod.rs
@@ -1,6 +1,3 @@
-mod checks;
-pub use checks::*;
-
 mod instruction_in_text;
 pub use instruction_in_text::*;
